☑️ Add development env auto-migration to test env
Note: I grabbed [code from Hyrax][1] which was introduced in
[this commit][2] via [this PR][3].

Prior to this commit, in the CI environment we would see the following
error:

```
Migrations are pending. To resolve this issue, run:

       bin/rails db:migrate RAILS_ENV=development
```

We had run the migrations for test, and later assert that the test
schema is up to date; however, there's an interplay between the Rails
'test' env and 'development'.  That interplay is an assumption in the
Rails 'test' environment that we're running tests in a similar
place/state as our development environment.  In other words, when we
boot rails in test we check the migrations of both test and development.

In CI, we don't have a development environment we've worked against, so
typically the migrations have not run.

With this commit, we run the development migrations that are pending.
Then we call the assert test schema.

The result is that we remove the reported error, because we've performed
the development migrations.

A side benefit to this is that whenever we run tests, we also will
migrate our development database.  Which addresses a few scenarios we
might miss as developers.

To verify that this does not obliterate development data, I ran the
following steps:

- `rails db:create db:migrate` :: Ensure we have a development
  environment.
- `rails db:seed` :: Load data into the development environment
- `rails runner "puts User.count"` :: To note the original
  user count.
- `rspec` :: run the tests (and therefore the added code about
  syncing the development migrations).
- `rails runner "puts User.count"` :: Check the user count has
  not changed.

[1]:https://github.com/samvera/hyrax/blob/b8c4fa4c8fddbb4d4d4b89fc4b514bd6d5d83928/spec/spec_helper.rb#L30-L36
[2]:https://github.com/samvera/hyrax/commit/5e09a0345f
[3]:https://github.com/samvera/hyrax/pull/4536