🎁 Add questions.data field
Given that the application is a repository for shareable questions, and
not responsible for prompting users to answer those questions, I don't
think we need worry about the complex modeling.

Instead each question will have a [serializer][1] that is responsible
for loading and dumping the data in an adequate format for the
application logic; namely importing from CSV (however we plan to do
that) and exporting to XML (however we plan to do that).

Related to:

- https://github.com/scientist-softserv/viva/issues/30

[1]: https://api.rubyonrails.org/classes/ActiveRecord/AttributeMethods/Serialization/ClassMethods.html#method-i-serialize