🎁 Add matching question data with serialization
We're recommending that we receive CSVs for ingesting questions.  As
such, we want a way to represent a single cell that would have one or
more pairs for matching.

This is a relatively simple encoding.  My thoughts being that `|` is a
good representative for multi-value.  But we need two different bits of
pairing logic.  Enter the `::` as the other.

Therefore `A::B|C::D` becomes `[["A", "B"] ["C", "D"]]`.

Related to:

- https://github.com/scientist-softserv/viva/issues/30