🎁 Adding serialization for drag and drop
In reviewing the types of drag and drop questions, there are two that
I've found:

1. Drag the ones that are correct into a drop zone
2. Drag answers to their correct drop zones

This commit adds functionality to handle each of those types.  It's
possible we'll want to consider extracting a new type to ease the
convolution of two types of validations; but for now the specs work and
extraction should be relatively straight forward.

Related to:
- https://github.com/scientist-softserv/viva/issues/30