🎁 Add Question::BowTie
The client provided example of Bow Tie does not match the external
material regarding what a Bow Tie question looks like.  As provided, the
Bow Tie example looks more like a Drag and Drop with a label for the
center value.

The Question::BowTie created follows the format outlined here:
https://allnurses.com/next-generation-nclex-what-t738934/

The reason I chose to implement towards the Bow Tie of the external
resource is that this is more likely to be QML compliant.

Related to:

- https://github.com/scientist-softserv/viva/issues/30