🎁 Add keyword_names and category_names to Question
There are three somewhat inter-related things happening in this commit:

1. Amending `.gitignore`
2. Updating factories to include auto generation of keywords and
   categories.
3. Adding the `Question.filter_as_json` method

In regards to the `.gitignore`, I created a local override file and
don't want to commit that into the repository.  The override was so that
I could use PGAdmin to run SQL queries to test the `Question.filter`
method.  ([see playbook article][1])

In regards to the factories, I wanted to have quick ways to
auto-generate keywords and categories.

Last, the `Question.filter_as_json`.  This is the guts of the change,
namely that the React components are looking to consume the following
attributes for each question returned via the search result: `id`,
`type`, `text`, `keyword_names`, and `category_names`

[1]:  https://playbook-staging.notch8.com/en/dev/database/using-pgadmin-with-docker