🎁 Expose Question#type_name for better legibility
With this commit, I introduce the Question#type_name that is intended
for use as a more human friendly format for chosing a type in the UI.

I have made attempts to not break the upstream work, but as that is in
flight, I may have broken things.

This builds on https://github.com/scientist-softserv/viva/pull/96 adding
the concept of type_name to the mix of type_label.