🎁 Expose Question::StimulusCaseStudy#data
Prior to this commit, the `Question::StimulusCaseStudy#data` was nil.  In
the database we still store that as `nil`.  But, with this commit we now
leverage the `Question::StimulusCaseStudy#child_questions` to build the
data.

This is then exposed when we render the results of
`Question#filter_as_json`.

Related to:

- https://github.com/scientist-softserv/viva/issues/63