🎁 Add Keyword and Category in CSV
Prior to this commit, we were not able to assign categories nor keywords
to questions during the import of a CSV.

With this commit, we now can.  See the specs for details on the column
structure.

Also, this includes a minor tweak to how we account for the
`Question.filter` of the given type_name(s); namely handling the Single
Table Inheritance paradigm of ActiveRecord (e.g. child and grand-child
or grand-grand child of a Question class).

Related to:

- https://github.com/scientist-softserv/viva/issues/86