♻️ Adjust `Question::Matching` to have many "right"
Prior to this commit, a matching question's "pairs" were a one to one
relation.  One value on the left and one value on the right.

With this commit, a matching question's "pairs" can have a one to many
relation; one value on the left and many values on the right.

*Note:* We are not looking for uniqueness of all of the values;
something that could be desirous; then again maybe a "right" could
belong to more than one "left".

After merging this PR, you will want to run `rails db:seed` to adjust
your existing questions to reflect the data changes.

Related to:

- https://github.com/scientist-softserv/viva/issues/58