♻️ Favor single method instead of method chain
In favoring a single method, we create a canonical source for the
questions.

In addition, by creating a class method call `.names` we can leverage
that on `Question#keywords#names` and `Question#categories#names` (a
helpful feature of Rail's query composer).

In addition, we also expose Level controlled vocabulary

This moves away from hard-coding; in that there's now a method that
defines the `Level.names`.  Then in the future, we can swap out how that
`Level.names` list is generated.

This builds from the conversation with the client.