🎁 Add CSV Header Verification
Prior to this commit, we were relying on the rows to check for valid
headers.

With this commit, we check for valid headers on the CSV; when there
aren't valid headers we have one error at the top level.

We also return a singular question to conform to the interface (though
that is up for discussion).