♻️ Refactor column validation to common method
Each CSV has a baseline set of headers it needs.  That validation is
similar to what must happen for each Question subtype.

This refactor creates the space for each Question subtype to specify the
expected headers (via the `.required_csv_headers` class_attribute).

That change will come after this refactor.