♻️ Create Traditional::CsvRow structure
The goal of this inner class is to provide validation of the CSV's
structure.  The refactor demonstrates that the interface has been
implemented.

The reason for validating the CSV is in #184; namely the validations at
the time of this commit are focused on the data being serialized.  The
current error messages speak about "hashes and arrays" a concept that is
not in the CSV.

The goal is to have more explicit messages about what is wrong with the
CSV; that is the next commit in this work.

Related to:

- https://github.com/scientist-softserv/viva/issues/184