♻️ Improve error messages for Traditional CSV import
Prior to this commit, the error messages mentioned the data should be
hashes and arrays.  This is confusing for the person importing the CSV
as they are thinking about rows and columns.

With this commit, the error message now speaks to the CSV's structure.
There is still error handling of data serialization.

Related:

- https://github.com/scientist-softserv/viva/issues/184