🎁 Add Traditional Question XML format
This is a "proof of concept" for exporting XML for a
`Question::Traditional` object.  The proof of concept should also work
for the `Question::SelectAllThatApply` (with some modifications).

I went with the absolute minimum setup to get an XML fragment generated.
There would be more as we look to generalize the behavior.

At present, there are no QA tests for exporting these entries.  In part
because only the "Traditional" question has an XML exporter.

Related to:

- https://github.com/scientist-softserv/viva/issues/173