♻️ Extract abstract ImportCsvRow class
Prior to this commit, we only had the concept of a
`Question::Traditional::ImportCsvRow`; it was a concrete implementation.

With this commit, I'm extracting an abstract class (e.g.
`Question::ImportCsvRow`) that can be extended by other Question
subclasses.  This refactor simplifies the work for adding validation to
`Question::SelectAllThatApply`.  See #181.

Related to:

- https://github.com/scientist-softserv/viva/185
- https://github.com/scientist-softserv/viva/184
- https://github.com/scientist-softserv/viva/183
- https://github.com/scientist-softserv/viva/182
- https://github.com/scientist-softserv/viva/181
- https://github.com/scientist-softserv/viva/180
- https://github.com/scientist-softserv/viva/164