♻️ Improve Question::ImportCsvRow interface
This commit does three things:

1. Indicates the abstract nature of `Question::ImportCsvRow`
2. Removes attr_reader declarations for unused variables (at the
   abstract level).
3. Adds a `Question::ImportCsvRow#valid?` method that does double
   duty.

The outstanding issue, for the moment, is the identified TODO item.
This is one that I think we will likely want, but is not something I'm
going to spend cycles on in this moment.  The impact of this is shifting
away from raising an exception in the UI if the ImportCsvRow succeeds at
validation but the underlying question fails.  Instead we will say it is
invalid but there will likely be no errors.  Not ideal, hence the TODO
item.