🎁 Add XML output for `Question::SelectAllThatApply`
Before this commit, we did not have a QTI export for
`Question::SelectAllThatApply`.  My goal is to get this export
information ready for the client to test exporting and loading into
Canvas.

This commit leverages work from #190 by repurposing a common view and
adding appropriate information.  Yes there's duplication, but until I
find a 3rd use case, I'm going to keep using that duplication.

Related to:

- https://github.com/scientist-softserv/viva/pull/190
- https://github.com/scientist-softserv/viva/pull/189
- https://github.com/scientist-softserv/viva/issues/171
- https://github.com/scientist-softserv/viva/issues/167
- https://github.com/scientist-softserv/viva/issues/166