🎁 Add StimulusCaseStudy Scenario Handling
This commit works with the easiest logic for sub-questions of a
`Question::StimulusCaseStudy`; namely the `Question::Scenario`.  The
`Question::Scenario` does not exist as a stand-alone question.
So this begins the process of building a case study.

Next up is to incorporate the case study for other sub-types.

Related to:

- https://github.com/scientist-softserv/viva/issues/197