🧹 Improve code-coverage
In reviewing the code coverage, there were places that would likely
never be hit but are important for defined interfaces or for catching
unlikely conditions or for initial application code.

There was also a vestigal method (e.g. `#search_params`) that could be
removed.

Further, I refactored a test to bring it inline with others and added a
test condition that would call previously untested code.

The result of this change is:

- moving code coverage from 98.99% to 99.45%
- testing an untested area of code
- improving an error message
- removing an unnecessary method