♻️ Favor `TEXT_` over SECTION_
Prior to this commit, we had `TEXT` and `SECTION_` column names for markdown
questions.  However, in conversation with team members the SECTION_ was
a bit more opaque.

What this change does is coerce all `TEXT` and `TEXT_` into the
`questions.text` column.

Up next is to coerce that text (which we're treating as Markdown) into
HTML and store that in the data.