🎁 Add Essay question
The essay question is complicated in that there is lots of text.  To
represent this in CSV requires some antics; namely treating the SECTION_
columns as lines for rendering.

What remains is considering whether or not to convert the text to
markdown.  For now we can defer on that decision; perhaps pairing with
the React developers who continue moving this forward.

Related to:

- https://github.com/scientist-softserv/viva/issues/179