📚 Include example files for import/export
On <2023-12-07 Thu>, I was scheduled to meet with a client to go over
our existing export.  I sent them two questions from the current state
of the export.  They also sent me exports from Canvas.  Comparing the
two I found significant differences; irreconcilable in my assessment.

This commit includes the client provide files (in the
`./examples/canvas-export` directory).

I then went to a [page to generate a CSV][1] that I would then convert
to [QTI via this page][2].  That is the contents of
`./examples/canconvert.k-state.edu/`.

The examples sent to the client are not included in this commit, as they
are to far afield from the `canvas-export` files and close but not quite
adequate to what was produced via [the converter][2].

Why not use the converter for all?  Because the converter only handles
`Question::SelectAllThatApply` and `Question::Traditional` type CSVs.

[1]: https://dl.sps.northwestern.edu/canvas/2021/06/add-quiz-questions-to-canvas-by-converting-csv-files-to-qti-zip-files/

[2]: http://canconvert.k-state.edu/qti/

Related to:

- https://github.com/scientist-softserv/viva/issues/166
- https://github.com/scientist-softserv/viva/issues/170
- https://github.com/scientist-softserv/viva/issues/171
- https://github.com/scientist-softserv/viva/issues/173