♻️ Create XML exports based on successful import
This commit got a bit ahead of me; namely it involves a few different
refactors.

1. XML output that maps to what worked in test.
2. Refining method names to be descriptive
3. Adding a few helper methods to ensure uniqueness.
4. Each question type will have it's own XML file.

A "vanity" refactor was moving from the `Question#to_xml` method to a
more conventional Rails view methodology.  The primary purpose for that
was to allow for the usage of `<%-`.  That idiom is a Rails specific
rendering specific feature for ERB.  The `<%-` means to not count that
line as output.  I use it quite a bit; and when using the
`Question#to_xml` approach it would produce a parse error.

Related to:

- https://github.com/scientist-softserv/viva/issues/173
- https://github.com/scientist-softserv/viva/issues/171
- https://github.com/scientist-softserv/viva/issues/166