🎁 Add QTI XML export for Categorization question
This builds on the verification of the <2023-12-11 Mon> client
verification of exporting the Matching question.

It is a naive refactor that circumvents the variance between Matching
and Categorization; namely that Matching has a cardinality of one and
Categorization has a cardinality of multiple.

The circumvention is written up in the following ticket:

- https://github.com/scientist-softserv/viva/issues/237

To address that issue, I'll need to work on a refactoring of the
matching and category CSV format.  I've explained that to Lisa and she's
aware of those considerations.

Related to:

- https://github.com/scientist-softserv/viva/issues/166
- https://github.com/scientist-softserv/viva/issues/170
- https://github.com/scientist-softserv/viva/issues/236