♻️ Add cardinality and unique validation to Matching behavior
Prior to this commit Matching and Categorization had the same import
logic.  However, the reality is that Matching allows only one to one
answer/choice pairs.  Where as Categorization allows multiple correct
elements per answer.

With this commit we introduce uniquness validation as well as
cardinality validation; as defined in the specs.

There is still a consideration of naming.  The `LEFT_` and `RIGHT_`
columns map to the bowtie concepts; but that's a digression.

Closes #237

See:

 - https://github.com/scientist-softserv/viva/issues/237