♻️ Improve error messaging around malformed CSV
We're dealing with character encoding issues; one almost certainly based
on Excel's default export assumptions.

This PR does not attempt to coerce the file from one encoding to
another, but instead favors adding a bit more information to the error
handling.  Ultimately pushing the problem back to the client.

A future change will be to attempt to coerce malformed text from one
format to another.  But that is a future issue.