🎁 View all bookmarks
This commit will add a way for the user to view all the bookmarked
questions.  My reasoning for the mix of a tags and button element is
because the first two buttons in the bookmarks dropdown should be anchor
tags since they are navigation elements and the third button, which is
the delete all bookmarks button, should stay a button because it is an
action element.

Ref:
  - https://github.com/scientist-softserv/viva/issues/281