🐛 Ensures unselected levels will post nil
This commit fixes an error where if a level was not selected by the user
when creating an essay question, a string or an incompatible Ruby
datatype (null) was being posted.
Now, if a level is not selected by the user, an empty string is passed
in the request and converted to nil in the controller.

Ref:
- https://github.com/scientist-softserv/viva/issues/294