🐛 Subject and Keyword duplication bug
This commit fixes a bug that allowed the user to add a subject or
keyword with a capitalized letter then add the same word with a
lowercase letter. As long as the capitalized word was added first, the
code logic would allow the lowercase word to be added as well. The fix
maps through the subject and keyword arrays and downcases each item
before checking for duplicates.