🎁 Add logic to the Question model for pg_search
This commit will add logic to the Question model to make the pg_search
gem work.  We have the new `searchable` field which we are indexing the
`text` and `data` field words onto it.  We had to pay special
consideration for HTML strings and also specifically how it would work
with the StimulusCaseStudy question.  For that, we want to index all of
its child works' texts and data onto the `searchable` field.