♻️ Refactor BookmarkExportService
This commit will add the `is_file` attribute onto the reformatters so we
can define all the settings from the service instead.  Now when it goes
into the BookmarkExportService, it will already have all the information
needed to run.  This allows us to do a major refactor and get rid of the
send method pattern.

Also, we changed `application/xml` to `text/xml` since that is more
appropriate.