✅ End-to-end test for user authentication
This commit adds Cypress testing for the user authentication workflow.
It ensures the user can log in, sign up, and log out as well as request
a password reset, reset their password, use the "remember me" check box,
and logged in users can access authenticated routes.

The current functionality of the user auth workflow is predominantly
happy path. Once these features exist in the application, the testing
should be updated to include:
- logged out users attempting to access authenticated routes and being
routed to the log in page (currently there is a Rails routing error)
- users attempting to sign in with invalid credentials should be given
feedback on their attempt (currently the user is just routed back to the
splash page)

Ref:
- https://github.com/notch8/viva/issues/409