🎁 Add question description field
This commit will add a question description field for the Essay and
Upload question type forms.  Previously, we were only able to set the
question text and it gets parsed into the question data field in the
record.  Now we can set them independently.  We changed the QuestionText
component to accept new props so we can customize the various texts that
are displayed.

Ref:
- https://github.com/notch8/viva/issues/431