🐛 Dismissing a filter resets the URL to root
This commit fixes a bug where removing one filter option would trigger a
rerouting of the page to the root directory. This was caused by
conflicts in state management between the Search/index, SearchBar, and
SearchFilters components. The fix involved a refactor to that moved
state and form submission handling to the Search/index component, and
passing down the necessary props to the SearchBar and SearchFilters
components.

Ref:
- https://github.com/notch8/viva/issues/436