🎁 Canvas exports to New and Classic platforms
This commit updates the CanvasService to handle exports to the New and
Classic platforms. This required a change to the imsmanifest.xml file
that adds additional attributes and metadata.

The ExportModal component was updated to note the difference between the
New and Classic platforms and maintain styling in a mid-size screen.

Ref:
- https://github.com/notch8/viva/issues/400